Restcontroller with admin managemant

+++ src/main/java/com/vn/jav/henllo/Repository/RoomTypeRepository.java
@@ -2,6 +2,8 @@ package com.vn.jav.henllo.Repository;
 
 import com.vn.jav.henllo.Model.RoomType;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 
-public interface RoomTypeRepository extends JpaRepository<RoomType,Long> {
+public interface RoomTypeRepository extends JpaRepository<RoomType,Integer> {
+
 }
